+++ Source Files/Vector2D.cpp
@@ -1,19 +1,19 @@
 #include "..\Header Files\Vector2D.h"
 
-Vector2D::Vector2D() : x_{ 0 }, y_{0}
+Vector2D::Vector2D() : x_{0.f}, y_{0.f}
 {
 
 }
 
-Vector2D::Vector2D(int x, int y): x_{x}, y_{y}
+Vector2D::Vector2D(float x, float y): x_{x}, y_{y}
 {
 
 }
 
 void Vector2D::zero()
 {
-	x_ = 0;
-	y_ = 0;
+	x_ = 0.f;
+	y_ = 0.f;
 }
 
 Vector2D& Vector2D::operator+(const Vector2D& other_v)
